Maintain: CatfishTest is added. Test for cleaners are moved to service package.
