+++ src/test/java/org/dice_research/opal/catfish/AllTests.java
@@ -1,6 +1,9 @@
 package org.dice_research.opal.catfish;
 
-import org.dice_research.opal.catfish.service.impl.ThemeCleaner;
+import org.dice_research.opal.catfish.service.EmptyBlankNodeCleanerTest;
+import org.dice_research.opal.catfish.service.FormatCleanerTest;
+import org.dice_research.opal.catfish.service.LiteralCleanerTest;
+import org.dice_research.opal.catfish.service.ThemeCleanerTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -14,7 +17,9 @@ import org.junit.runners.Suite.SuiteClasses;
 
 		ThemeCleanerTest.class,
 
-		LiteralCleanerTest.class
+		LiteralCleanerTest.class,
+
+		CatfishTest.class
 
 })
 public class AllTests {
